test(conda-types): drop the deterministic formatting matrix
The proptests cover the same ground (canonical round-trips, condition
ASTs, panic hunting) with random inputs and committed regression seeds,
so the 744-line hand-built matrix added little on top. Its one unique
piece, the legacy Display scalar-quoting regression guard, moves into
the parse.rs tests.
